enable sbt 2 wartremoverTask
