+++ sbt-plugin/src/main/scala-3/wartremover/WartRemoverCompat.scala
@@ -1,7 +1,5 @@
 package wartremover
 
-import sbt.Configuration
-import sbt.Def
 import java.io.File
 import xsbti.FileConverter
 
@@ -9,5 +7,14 @@ private[wartremover] trait WartRemoverCompat { self: WartRemover.type =>
   private[wartremover] def convertToFile(x: xsbti.VirtualFileRef, fileConverter: FileConverter): File =
     fileConverter.toPath(x).toFile
 
-  private[wartremover] def wartremoverTaskSetting(x: Configuration): Seq[Def.Setting[?]] = Nil
+  private[wartremover] def packageBinCopy(taskName: String, outputFileName: String): String =
+    s"""
+       |TaskKey[Unit]("${taskName}") := Def.uncached(
+       |  IO.copyFile(
+       |    fileConverter.value.toPath((Compile / packageBin).value).toFile,
+       |    file("${outputFileName}")
+       |  )
+       |)
+       |""".stripMargin
+
 }
