+++ sbt-plugin/src/main/scala/wartremover/WartRemover.scala
@@ -9,12 +9,16 @@ import sjsonnew.JsonFormat
 import sjsonnew.support.scalajson.unsafe.CompactPrinter
 import java.io.FileInputStream
 import java.util.zip.ZipInputStream
+import java.net.URLClassLoader
+import java.lang.reflect.Modifier
+import scala.annotation.tailrec
 import scala.collection.concurrent.TrieMap
 import scala.concurrent.Await
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
 import scala.concurrent.duration.DurationInt
 import scala.util.Using
+import wartremover.InspectWart.Type
 // format: off
 import sbt.{given, *}
 // format: on
@@ -300,14 +304,16 @@ object WartRemover extends sbt.AutoPlugin with WartRemoverCompat {
     val res = wartRunCache.getOrElseUpdate(
       key,
       Future {
+        val outputFileName = "result.jar"
+        val taskName = "createJar"
         val buildSbt =
           s"""
              |logLevel := Level.Warn
              |scalaVersion := "${key.scalaV}"
-             |crossPaths := false
              |libraryDependencies := Seq(
              |  "org.wartremover" %% "wartremover" % "${Wart.PluginVersion}"
              |)
+             |${packageBinCopy(taskName = taskName, outputFileName = outputFileName)}
              |""".stripMargin
 
         IO.withTemporaryDirectory { dir =>
@@ -321,13 +327,12 @@ object WartRemover extends sbt.AutoPlugin with WartRemoverCompat {
             Seq(
               "-jar",
               launcher.getCanonicalPath,
-              packageBin.key.label
+              taskName
             )
           )
           assert(exitCode == 0, s"exit code = $exitCode")
           log.info(s"compiled ${key.files.size} files")
-          val Seq(compiledJar) = (dir / "target").listFiles(f => f.isFile && f.getName.endsWith(".jar")).toList
-          IO.readBytes(compiledJar).toSeq
+          IO.readBytes(dir / outputFileName).toSeq
         }
       }(using ExecutionContext.global)
     )
@@ -577,4 +582,88 @@ object WartRemover extends sbt.AutoPlugin with WartRemoverCompat {
 
   private[wartremover] def derive[T](s: Setting[T]): Setting[T] =
     Def derive s
+
+  private def wartremoverTaskSetting(x: Configuration): Seq[Def.Setting[?]] = {
+    x / wartremoverTask := Def.inputTaskDyn {
+      val parsed0 = InspectArgsParser.get(file(".").getAbsoluteFile.toPath).parsed
+      val parsed = InspectArgs.from(parsed0.collect { case x: InspectArg.Wart =>
+        x
+      })
+      Def.taskDyn {
+        val errResult = compileWartFromSources(parsed(Type.Err)).value
+        val warnResult = compileWartFromSources(parsed(Type.Warn)).value
+        createInspectTask(
+          x = x,
+          warningWartNames = parsed(Type.Warn).warts ++ warnResult.wartNames,
+          errorWartNames = parsed(Type.Err).warts ++ errResult.wartNames,
+          jarFiles = errResult.jarBinary.toSeq ++ warnResult.jarBinary.toSeq,
+        )
+      }
+    }.evaluated
+  }
+
+  private def compileWartFromSources(inspectArg: InspectArgs): Def.Initialize[Task[CompileResult]] = {
+    Def.taskDyn {
+      val log = streams.value.log
+      val wartremoverJars = Def.taskDyn {
+        val wartremoverCross = wartremoverCrossVersion.value
+        getJarFiles(
+          "org.wartremover" %% "wartremover" % Wart.PluginVersion cross wartremoverCross
+        )
+      }.value
+      if (inspectArg.sources.nonEmpty) {
+        val bytes: Seq[Byte] = getJar(inspectArg.sources.toSet).value
+        IO.withTemporaryDirectory { tmpDir =>
+          val compiledJar = tmpDir / "wartremover" / "warts.jar"
+          IO.write(compiledJar, bytes.toArray)
+          log.debug(s"compiled jar = $compiledJar")
+          val classes = getAllClassNamesInJar(compiledJar)
+          if (classes.isEmpty) {
+            log.error("not found compiled classes")
+            Def.task(CompileResult.empty)
+          } else {
+            log.info(s"compiled classes = ${classes.mkString(" ")}")
+
+            val xs = Using.resource(
+              new URLClassLoader(
+                (compiledJar.toURI.toURL +: wartremoverJars.map(_.toURI.toURL)).toArray
+              )
+            ) { loader =>
+              classes.filter { className =>
+                val clazz = Class.forName(className, false, loader)
+                val traverserName = "org.wartremover.WartTraverser"
+
+                @tailrec
+                def loop(c: Class[?]): Boolean = {
+                  if (c == null) {
+                    false
+                  } else if (c.getName == traverserName) {
+                    true
+                  } else {
+                    loop(c.getSuperclass)
+                  }
+                }
+
+                if (Modifier.isAbstract(clazz.getModifiers)) {
+                  false
+                } else {
+                  loop(clazz)
+                }
+              }.map { s =>
+                if (s.endsWith("$")) s.dropRight(1) else s
+              }
+            }
+
+            Def.task(
+              CompileResult(jarBinary = Some(bytes), wartNames = xs.map(Wart.custom))
+            )
+          }
+        }
+      } else {
+        Def.task(
+          CompileResult.empty
+        )
+      }
+    }
+  }
 }
